requirements.txt
Flask
python-dotenv
requests
openai   # or groq sdk if using Groq
